Initial Implementation of Python Library
This is the initial submission of the ``smis`` python library, which sets the project and python modules, as well as, the related documentation.

In this submission, a ``Client`` class is implemented that provides the entry point interface to access the service. An internal ``_Resource`` class is provided to represents service resources.

Initial documentation for the classes and for the project is provided, so that we can begin the integration with Read-The-Docs.