Implemented Authentication Proxy
This submission implements an ``OxygenAuthenticationProxy`` class that takes care of authorizing an application and a user that has access to the MIS service. The class requires client key and secret pairs to authenticate the app, and a ``login_callback`` function that client applications can use to log in a user.