Updated Required Modules to Latest Versions
This change updates the required dependencies to the latest versions to keep up with defect fixes and new functionality. While updating the dependencies, I've found a change in the way ``requests-oauthlib`` handles the process of fetching the access token, which requires the ``oauth_verifier`` to be specified. Unfortunately, Oxygen does not provide a verifier, so after looking at possible alternatives, I realized that I needed to use the ``OAuth1`` helper, which does not require the verifier, instead of ``OAuth1Session``.

The change is in the implementation of ``OxygenAuthenticationProxy._access_token()``, which should not affect client code. The code is a little messier, but it allows me to use the latest version of ``requests-oauthlib``, and I was behind quite a few versions, so the update was needed.