Restructured Project for Python Only Library
After much thought, I decided that this project will be Python only and that additional libraries for other languages will be provided in separate projects and repos. This allows me to make the structure of the project python specific and remove the ``proj/python`` structure bringing ``source`` and ``tests`` to the root of the project. Changes to the documentation build have been made to support the new structure. I also consolidated the required packages files into a single ``requirements.txt`` which matches what other python projects are doing.