Versioned Library and Documentation
After playing with different options to automate the process of versioning the library, I realize that it is very difficult to maintain a consistent version between source code and documentation because we are using two different systems hooked to GIT to build and test the library and to generate the documentation. Furthermore, the inclusion of a build number doesn't seem necessary, and to manage releases can be done at the begining of changes when a release is planned and the version can be updated in the library itself.

These changes insure that the project information and version is consistent between the library and the documentation.