Modifying Interface to Return Python Objects from ``get()`` Methods
This change modifies the ``get()`` methods of ``Client`` and ``_Resource`` to return the resource payload as a python object instead of the straight response object. This simplifies the interface where client code can manipulate the result from a ``get()`` call instead of having to grab the payload from the response object.

The classes also provide a ``response`` property which returns the latest ``Response`` from the service. The properties will return ``None`` if ``get()`` has not been invoked.

Testing Performed
-----------------

Unit tests have been provided or updated to reflect the new behavior. Mocks have been moved to a new ``mocks.py`` module to consolidate where test mocks are implemented. The acceptance harness has been modified to expect the new behavior and all tests pass.